add cwl tool definition for alignment docker image
this doesn't *quite* work with the reference cwl implementation bug it's
a good start

outstanding issues:

1) the reference implementation mounts input paths into the container as
read-only volumes. the pipeline by default writes logs etc. to the same
path as the bam file by default, which bombs out, because that dir is
mounted read-only. ideally we would pass the "designated output
directory" as the place to store logs and scratch files. i started a
thread about this:
https://groups.google.com/forum/#!topic/common-workflow-language/hZetEsClF9Y. for
now working about by changing the implementation to mount everything
read/write, not a good long-term solution.

2) related to the above, the output globbing works against the
designated output directory, which by default is the current working
directory. in order for the output globs to work, the output directroy
has to be specified to match the directory the bamfiles are in. this
can't be passed as part of the job, it's just a command line argument to
cwltool, so this makes the whole process very brittle and not
reproducible. again, the ideal thing would be if i could get access to
the outdir in the workflow definition via an Expression and pass that as
-l and -s to the pipeline.